[WIP] Transitive HeaderMaps Prov, hdr_mapping attr
This patch adds an explicit header mapping to map namespaces to headers.

A given `objc_library` enters a header mapping. At the end of the day, this is
propagated to a single headermap.

We make an additional headermap for all of the deps.

Now, there is 2 header maps:

1) an iquoted headermap of all the internal sources
2) an included headermap for all of the deps

This is the most efficient scheme for this project.

TODO:
- Remove HeaderNamespace
- Find a better way to get `hdrs`